Revert "Refactoring Point #1"
